Move non-unit export from units.ts to unit-registry.ts
